% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-edge-circular.R
\name{annotate_arc_rect}
\alias{annotate_arc_rect}
\title{Special annotate layers}
\usage{
annotate_arc_rect(start = 180, end = 360, r0 = 0.6, r1 = 1, ...)
}
\arguments{
\item{start}{start angle in degree.}

\item{end}{end angle in degree.}

\item{r0}{inner radius of arc-rect.}

\item{r1}{outer radius of arc-rect.}

\item{...}{other parameters passing to `geom_polygon()`.}
}
\value{
a layer object.
}
\description{
Special layers to add arc-rect for circular layout.
}
\author{
Hou Yun
}
