% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-special.R
\name{as_bipartite_circular}
\alias{as_bipartite_circular}
\title{Special layout}
\usage{
as_bipartite_circular(
  g,
  inner_nodes = NULL,
  outer_nodes = NULL,
  start = 0,
  end = 180,
  r0 = 0.6,
  r1 = 1,
  ...
)
}
\arguments{
\item{g}{a igraph object.}

\item{inner_nodes, outer_nodes}{character vector indicating which nodes are
internal and which are external.}

\item{start}{start angle in degree.}

\item{end}{end angle in degree.}

\item{r0}{inner radius of arc-rect.}

\item{r1}{outer radius of arc-rect.}

\item{...}{other parameters passing to `igraph::as.igraph()`.}
}
\value{
a igraph object.
}
\description{
Special layout of arranging node on a half-circle.
}
\author{
Hou Yun
}
