% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{z_score}
\alias{z_score}
\alias{p_score}
\alias{centrality_in_module}
\alias{centrality_between_module}
\title{Network Attributions}
\usage{
z_score(g, module_vars = NULL, ...)

p_score(g, module_vars = NULL, degree_vars = NULL, ...)

centrality_in_module(module_vars = NULL, ...)

centrality_between_module(module_vars = NULL, degree_vars = NULL, ...)
}
\arguments{
\item{g}{a igraph object.}

\item{module_vars}{if NULL (default), will uses \code{igraph::cluster_fast_greedy()}
function to calculate module index.}

\item{...}{other parameters passing to \code{igraph::cluster_fast_greedy()}.}

\item{degree_vars}{if NULL (default), will uses \code{igraph::degree()} to calculate
degree attributes of nodes.}
}
\value{
a numeric vectors.
}
\description{
helper functions to get nodes/edges/graph attributions.
}
\author{
Hou Yun
}
