% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{geom_node_magic_text}
\alias{geom_node_magic_text}
\title{Magic labels of nodes}
\usage{
geom_node_magic_text(
  mapping = NULL,
  data = NULL,
  position = "identity",
  geom = "text",
  parse = FALSE,
  ...
)
}
\arguments{
\item{mapping}{set of aesthetic mappings created by \code{aes()} or \code{aes_()}.}

\item{data}{the data to be displayed in this layer.}

\item{position}{position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{geom}{one of text, label or richtext.}

\item{parse}{logical or a parse function. IF TRUE (default), the labels will
be parsed into expression.}

\item{...}{others passing to \code{ggplot2::geom_<geom>()}.}
}
\value{
a gg object.
}
\description{
layer function to draw node labels.
}
\author{
Hou Yun
}
