% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-special.R
\name{layout_bipartite_circular}
\alias{layout_bipartite_circular}
\title{Special layout}
\usage{
layout_bipartite_circular(g, ...)
}
\arguments{
\item{g}{a igraph object created by `as_bipartite_circular()`.}

\item{...}{not used.}
}
\value{
a layout matrix.
}
\description{
Special layout of arranging node on a half-circle.
}
\author{
Hou Yun
}
