% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-circle.R
\name{layout_within_circle}
\alias{layout_within_circle}
\alias{layout_on_circle}
\alias{layout_zoomin_circle}
\title{Circular Layout}
\usage{
layout_within_circle(
  g,
  layout = NULL,
  group = NULL,
  center = NULL,
  zoom = 0.9,
  by_nodes = FALSE,
  ...
)

layout_on_circle(g, group = NULL, center = NULL, zoom = 0.618, by_nodes = TRUE)

layout_zoomin_circle(g, layout = NULL, center = NULL, ...)
}
\arguments{
\item{g}{igraph object.}

\item{layout}{function used to calculate node positions.}

\item{group}{if NULL (default), all nodes will be treated as single group.}

\item{center}{a list named with x and y.}

\item{zoom}{numeric, range in [0, 1].}

\item{by_nodes}{logical, indicating whether to scale based on the number of nodes.}

\item{...}{other parameters passing to \code{layout} function.}
}
\value{
a two-columns matrix.
}
\description{
arranges nodes in circles according to group information.
}
\seealso{
Other layout: 
\code{\link{layout_packcircles}()}

Other layout: 
\code{\link{layout_packcircles}()}

Other layout: 
\code{\link{layout_packcircles}()}
}
\author{
Hou Yun
}
\concept{layout}
