% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-circular.R
\name{layout_in_circular}
\alias{layout_in_circular}
\alias{layout_on_circular}
\alias{layout_zoomin_circle}
\title{Circular Layout}
\usage{
layout_in_circular(
  g,
  layout = NULL,
  group_vars = NULL,
  center = NULL,
  zoom = 0.618,
  ...
)

layout_on_circular(g, group_vars = NULL, center = NULL, zoom = 0.618)

layout_zoomin_circle(g, layout = NULL, zoom = 0.97, ...)
}
\arguments{
\item{g}{igraph object.}

\item{layout}{function used to calculate node positions.}

\item{group_vars}{if NULL (default), all nodes will be treated as single group.}

\item{center}{a list named with x and y.}

\item{zoom}{numeric, range in [0, 1].}

\item{...}{other parameters passing to \code{layout} function.}
}
\value{
a two-columns matrix.
}
\description{
arranges nodes in circles according to group information.
}
\seealso{
Other layout: 
\code{\link{layout_fr_circular}()},
\code{\link{layout_packcircles}()}

Other layout: 
\code{\link{layout_fr_circular}()},
\code{\link{layout_packcircles}()}

Other layout: 
\code{\link{layout_fr_circular}()},
\code{\link{layout_packcircles}()}
}
\author{
Hou Yun
}
\concept{layout}
