% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-fr-circular.R
\name{layout_fr_circular}
\alias{layout_fr_circular}
\title{Circular Layout With FR}
\usage{
layout_fr_circular(g, group_vars = NULL, ...)
}
\arguments{
\item{g}{igraph object.}

\item{group_vars}{if NULL (default), all nodes will be treated as single group.}

\item{...}{other parameters passing to \code{layout_in_circular()} function.}
}
\value{
a two-columns matrix.
}
\description{
arranges nodes in circles and each group center is determine by
Fruchterman-Reingold layout.
}
\seealso{
Other layout: 
\code{\link{layout_in_circular}()},
\code{\link{layout_packcircles}()}
}
\author{
Hou Yun
}
\concept{layout}
