% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-liner.R
\name{layout_nice_liner}
\alias{layout_nice_liner}
\title{Liner layout}
\usage{
layout_nice_liner(g, ..., angle = 0)
}
\arguments{
\item{g}{a igraph object.}

\item{...}{condition of sort by, see `?dplyr::arrange` for details.}

\item{angle}{slope of line (in degree).}
}
\value{
a two-columns matrix of position.
}
\description{
this layout puts all nodes on a line.
}
\author{
Hou Yun
}
