% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-edge-circular.R, R/ggproto.R
\docType{data}
\name{netET-extensions}
\alias{netET-extensions}
\alias{GeomEdgeCircular}
\alias{GeomNodeTextCircular}
\title{netET package extensions to ggplot2}
\description{
In general the actual classes should be of little interest to users as the
standard ggplot2 api of using geom_* and stat_* functions for building up
the plot is encouraged.
}
\keyword{datasets}
