% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize matrix}
\usage{
normalize(mat)
}
\arguments{
\item{mat}{a layout matrix.}
}
\value{
a normalized matrix.
}
\description{
Normalize layout matrix in unit circle.
}
\author{
Hou Yun
}
