% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-otus.R
\name{select_otus}
\alias{select_otus}
\title{Select OTUs Based on Occurence and Abundance}
\usage{
select_otus(
  otu,
  otu_on_rows = TRUE,
  occur_thres = 0.5,
  abund_thres = 0.001,
  reverse = NULL
)
}
\arguments{
\item{otu}{OTUs table, should be a data.frame or matrix.}

\item{otu_on_rows}{logical, if TRUE means OTUs are on rows, and samples are
on columns.}

\item{occur_thres}{threshold value of occurence.}

\item{abund_thres}{threshold value of abundance.}

\item{reverse}{provides a indicator of positive or negative selection.}
}
\value{
OTUs table.
}
\description{
Select otus based on occurence and abundance.
}
\author{
Hou Yun
}
