% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhou_otu.R
\docType{data}
\name{zhou_otu}
\alias{zhou_otu}
\title{OTU table}
\format{
## `zhou_otu`
A data frame with 8816 rows and 60 columns:
\describe{
  row names of zhou_otu are OTU ID, and column names of zhou_otu are sample ID.
}
}
\usage{
zhou_otu
}
\description{
OTU table from Zhou(2022), see references for details.
}
\note{
Thanks to Dr. Zhou for providing and authorizing the original data.
}
\references{
Zhou HL, Liu XY, Zhou C, Han SF, Xu FR, Dong X. Root-Associated
Microbiomes of Panax notoginseng under the Combined Effect of Plant Development
and Alpinia officinarum Hance Essential Oil. Molecules. 2022 Sep 15;27(18):6014.
doi: 10.3390/molecules27186014. PMID: 36144749; PMCID: PMC9501277.
}
\keyword{datasets}
