% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhou_taxonomy.R
\docType{data}
\name{zhou_taxonomy}
\alias{zhou_taxonomy}
\title{Taxonomy table}
\format{
## `zhou_taxonomy`
A data frame with 8816 rows and 7 columns:
\describe{
  \item{Kingdom}{Kingdom of root-associated microbiomes.}
  \item{Phylum}{Phylum of root-associated microbiomes.}
  \item{Class}{Class of root-associated microbiomes.}
  \item{Order}{Order of root-associated microbiomes.}
  \item{Family}{Family of root-associated microbiomes.}
  \item{Genus}{Genus of root-associated microbiomes.}
  \item{Species}{Species of root-associated microbiomes.}
}
}
\usage{
zhou_taxonomy
}
\description{
Taxonomy table from Zhou(2022), see references for details.
}
\note{
Thanks to Dr. Zhou for providing and authorizing the original data.
}
\references{
Zhou HL, Liu XY, Zhou C, Han SF, Xu FR, Dong X. Root-Associated
Microbiomes of Panax notoginseng under the Combined Effect of Plant Development
and Alpinia officinarum Hance Essential Oil. Molecules. 2022 Sep 15;27(18):6014.
doi: 10.3390/molecules27186014. PMID: 36144749; PMCID: PMC9501277.
}
\keyword{datasets}
